pyproject.toml
[project]
name = "clipsync"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "streamlit>=1.31.0",
    "google-cloud-aiplatform>=1.38.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "faiss-cpu>=1.7.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]

[project.scripts]
api = "uvicorn backend.main:app --reload --host 0.0.0.0 --port 8000"
ui = "streamlit run frontend/app.py"
